requirements.txt
Flask
flask_cors
keras
tensorflow
keras
numpy
matplotlib
opencv-python
Pillow
colorama
requests
sendgrid
